added date entry and confirmations

+++ FlooringMaster/src/main/java/com/sg/flooringmaster/controller/FlooringMasterController.java
@@ -13,6 +13,7 @@ import com.sg.flooringmaster.dto.ProductNotFoundException;
 import com.sg.flooringmaster.dto.TaxRate;
 import com.sg.flooringmaster.service.FlooringMasterServiceLayer;
 import com.sg.flooringmaster.ui.FlooringMasterView;
+import com.sg.flooringmaster.ui.UserIONoValueException;
 import java.math.BigDecimal;
 import java.util.List;
 
@@ -75,41 +76,55 @@ public class FlooringMasterController {
     }
 
     public void editOrder() {
-        String date = view.getDate();
+        String date = null;
         List<Order> orderList = null;
         Order orderToEdit = null;
-
         Integer orderNumber;
 
         try {
+            date = view.getDate();
+
             orderList = service.getOrdersByDate(date);
             view.displayOrders(orderList);
 
             orderNumber = view.getOrderNumber(orderList);
             orderToEdit = service.getOrder(date, orderNumber.toString());
 
-        } catch (FlooringMasterNoOrdersForDateException e) {
-            view.displayErrorMessage("No orders found for date: " + date);
+            editTheOrder(date, orderToEdit);
+
+        } catch (FlooringMasterNoOrdersForDateException | UserIONoValueException e) {
+            System.out.println(e.getClass().getName());
+            if (e.getClass().getName().contains("FlooringMasterNoOrdersForDateException")) {
+                view.displayErrorMessage("No orders found for date: " + date);
+            }
         }
 
     }
 
     public void removeOrder() {
-        String date = view.getDate();
+        String date = null;
         List<Order> orderList = null;
         Order orderToRemove = null;
 
         Integer orderNumber;
 
         try {
+            date = view.getDate();
             orderList = service.getOrdersByDate(date);
             view.displayOrders(orderList);
 
             orderNumber = view.getOrderNumber(orderList);
-            orderToRemove = service.removeOrder(date, orderNumber.toString());
-
-        } catch (FlooringMasterNoOrdersForDateException e) {
-            view.displayErrorMessage("No orders found for date: " + date);
+            orderToRemove = service.getOrder(date, orderNumber.toString());
+            view.displayOrder(orderToRemove);
+            String confirmation = view.getConfirmation("Remove this order?");
+            if (confirmation.toLowerCase().equals("y")) {
+                orderToRemove = service.removeOrder(date, orderNumber.toString());
+            }
+        } catch (FlooringMasterNoOrdersForDateException | UserIONoValueException e) {
+            System.out.println(e.getClass().getName());
+            if (e.getClass().getName().contains("FlooringMasterNoOrdersForDateException")) {
+                view.displayErrorMessage("No orders found for date: " + date);
+            }
         }
 
     }
@@ -124,7 +139,7 @@ public class FlooringMasterController {
         Product product = null;
         TaxRate taxrate = null;
 
-        BigDecimal area;
+        BigDecimal area = null;
 
         String customerName = null;
         String prodType = null;
@@ -134,9 +149,29 @@ public class FlooringMasterController {
         List<TaxRate> taxRates = service.getAllTaxRates();
         List<Product> allProducts = service.getAllProducts();
 
-        customerName = view.getCustomerName();
+        invalidData = true;
+        while (invalidData) {
+            try {
+                date = view.getDate();
+                System.out.println("date: " + date);
+                invalidData = false;
+            } catch (UserIONoValueException e) {
+                view.displayErrorMessage("You must enter a value for date.");
+            }
+        }
+
+        invalidData = true;
+        while (invalidData) {
+            try {
+                customerName = view.getCustomerName();
+                invalidData = false;
+            } catch (UserIONoValueException e) {
+                view.displayErrorMessage("You must enter a value for name.");
+            }
+        }
 
         // get the product type
+        invalidData = true;
         while (invalidData) {
             try {
                 prodType = view.getProductTypeOption(allProducts);
@@ -159,25 +194,100 @@ public class FlooringMasterController {
             }
         }
 
-        area = view.getArea();
+        invalidData = true;
+        while (invalidData) {
+            area = view.getArea();
+            invalidData = false;
+        }
 
         Order newOrder = new Order(orderNumber, customerName, taxrate, product, area);
-        newOrder = service.addOrder(date, newOrder);
 
-        view.displayBanner("NewOrder:" + newOrder.toString());
+        view.displayOrder(newOrder);
 
-        //order = view.createNewOrder(taxRates, allProducts);
-        // List<Product> products                = Order newOrder = view.getNewOrder();
+        String confirmation = view.getConfirmation("Save this order?");
+        if (confirmation.toLowerCase().equals("y")) {
+
+            newOrder = service.addOrder(date, newOrder);
+
+            view.displayBanner("NewOrder:" + newOrder.toString());
+
+            //order = view.createNewOrder(taxRates, allProducts);
+            // List<Product> products                = Order newOrder = view.getNewOrder();
+        }
+
+    }
+
+    public void editTheOrder(String orderDate, Order order) {
+
+        Boolean invalidData = true;
+
+        Integer orderNumber = order.getOrderNumber();
+
+        Product product = null;
+        TaxRate taxrate = null;
+
+        BigDecimal area;
+
+        String customerName = null;
+        String prodType = null;
+        String state = null;
+        String editDate = "testAdddate";
+
+        List<TaxRate> taxRates = service.getAllTaxRates();
+        List<Product> allProducts = service.getAllProducts();
+
+        editDate = view.getEditDate(orderDate);
+
+        customerName = view.getEditCustomerName(order);
+
+        // get the product type
+        while (invalidData) {
+            try {
+                prodType = view.getEditProductTypeOption(order, allProducts);
+                product = service.getProduct(prodType);
+                invalidData = false;
+            } catch (ProductNotFoundException e) {
+                view.displayErrorMessage(e.getMessage());
+            }
+        }
+
+        // get the taxrate
+        invalidData = true;
+        while (invalidData) {
+            try {
+                state = view.getEditStateOption(order, taxRates);
+                taxrate = service.getTaxRate(state);
+                invalidData = false;
+            } catch (TaxRateNotFoundException e) {
+                view.displayErrorMessage(e.getMessage());
+            }
+        }
+
+        area = view.getEditArea(order);
+
+        Order newOrder = new Order(orderNumber, customerName, taxrate, product, area);
+
+        view.displayOrder(newOrder);
+
+        String confirmation = view.getConfirmation("Save changes?");
+        if (confirmation.toLowerCase().equals("y")) {
+            service.removeOrder(orderDate, orderNumber.toString());
+            newOrder = service.addOrder(editDate, newOrder);
+            // saveChanges();
+        }
+//        view.displayBanner("NewOrder: " + editDate + " | " + newOrder.toString());
     }
 
     public void displayOrders() {
         // display the orders for a specific date
-        String date = view.getDate();
+        String date;
         try {
+            date = view.getDate();
+
             List<Order> orderList = service.getOrdersByDate(date);
             view.displayOrders(orderList);
-        } catch (FlooringMasterNoOrdersForDateException e) {
-            view.displayErrorMessage("No orders found for date: " + date);
+        } catch (FlooringMasterNoOrdersForDateException | UserIONoValueException e) {
+            view.displayErrorMessage(e.getMessage());
         }
 
     }
